Abstract base `lttng_enabler` to support other types of enablers
Future commits will introduce the concept of event notifier enablers
which are similar to the event enablers already present with a few
differences.

To prepare the ground, this commit moves event dependent fields from the
`struct lttng_enabler` struct to a specialized `struct
lttng_event_enabler`.

The `lttng_enabler_*()` functions cannot be renamed because they are
currently exported. These are only used internally, so we can stop using
them and add new `lttng_event_enabler_*()` functions marked as hidden
which won't be exported (see liblttng-ust/ust-events-internal.h).
`lttng_enabler_*()` functions are kept in the ust-events.h file but
moved to the bottom of the file with other deprecated symbols.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I6d0af091c257a9ad297d17fca8993956279eb848